Move all user, auth, and session methods to different user.controller.ts.
